Flashcard Generator/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
transformers==4.12.3
torch==1.10.0
nltk==3.6.5
scikit-learn==0.24.2
PyYAML==5.4.1
datasets==1.15.1
evaluate==0.2.2
accelerate==0.5.1
